+++ Editor/EmacsSettings.cs
@@ -1,20 +1,11 @@
-using System.IO;
-using System;
-using UnityEditor.Callbacks;
 using UnityEditor;
-using UnityEngine;
-using UnityEngine.UIElements;
-using System.Linq;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using System.Collections.ObjectModel;
 
 namespace Emacs {
 
-  public class EmacsSettings : SettingsProvider {
+  public static class EmacsSettings {
     private const string _enabledName = "Emacs/Enable";
     private const string _emacsclientPathName = "Emacs/EmacsclientPath";
-    private const string _fileExtensionsName = "Emacs/FileExtensions";
+    private const string _fileMatchPatternName = "Emacs/FileMatchPattern";
 
 #if UNITY_EDITOR_WIN
     private const string _defaultEmacsclientPath =
@@ -26,8 +17,8 @@ namespace Emacs {
     private const string _defaultEmacsclientPath = "emacsclient";
 #endif
 
-    private const string _defaultFileExtensions =
-        ".cs,.txt,.js,.javascript,.json,.html,.shader,.template,.proto";
+    private const string _defaultFileMatchPattern =
+        "(.cs|.txt|.js|.javascript|.json|.html|.shader|.template|.proto|.xml)$";
 
     public static bool Enabled {
       get { return EditorPrefs.GetBool(_enabledName, false); }
@@ -39,33 +30,32 @@ namespace Emacs {
       set { EditorPrefs.SetString(_emacsclientPathName, value); }
     }
 
-    public static string FileExtensions {
-      get { return EditorPrefs.GetString(_fileExtensionsName, _defaultFileExtensions); }
-      set {
-        EditorPrefs.SetString(_fileExtensionsName, string.Join(",", SplitFileExtensions(value)));
+    public static string FileMatchPattern {
+      get { return EditorPrefs.GetString(_fileMatchPatternName, _defaultFileMatchPattern); }
+      set { EditorPrefs.SetString(_fileMatchPatternName, value); }
+    }
+
+#if UNITY_2018_3_OR_NEWER
+    private class EmacsSettingsProvider : SettingsProvider {
+      public EmacsSettingsProvider() : base("Preferences/Emacs", SettingsScope.User) {}
+
+      public override void OnGUI(string searchContext) {
+        OnGUI();
+      }
+
+      [SettingsProvider]
+      public static SettingsProvider CreateSettingsProvider() {
+        return new EmacsSettings();
       }
     }
-
-    public static IEnumerable<string> EnumerableFileExtensions {
-      get { return SplitFileExtensions(FileExtensions); }
-    }
-
-    private static IEnumerable<string> SplitFileExtensions(string extensions) {
-      return Regex.Split(extensions, @"\s*,\s*").Where(e => !string.IsNullOrEmpty(e));
-    }
-
-    // SettingsProvider
-    public EmacsSettings() : base("Preferences/Emacs", SettingsScope.User) {}
-
-    public override void OnGUI(string searchContext) {
+#else
+    [PreferenceItem("Emacs")]
+#endif
+    private static void OnGUI() {
       Enabled = EditorGUILayout.Toggle("Enabled", Enabled);
-      EmacsclientPath = EditorGUILayout.DelayedTextField("Emacsclient Path", EmacsclientPath);
-      FileExtensions = EditorGUILayout.DelayedTextField("File Extensions", FileExtensions);
-    }
-
-    [SettingsProvider]
-    public static SettingsProvider CreateSettingsProvider() {
-      return new EmacsSettings();
+      EmacsclientPath = EditorGUILayout.TextField("Emacsclient Path", EmacsclientPath);
+      FileMatchPattern = EditorGUILayout.TextField("File Extensions", FileMatchPattern);
     }
   }
+
 }