+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for Unity versions older than 2019.x.
+
+### Changed
+- Replaced filename extension preference with filename regex.
+- Fixed `OnOpenAssetAttribute()` error for Unity 2019.2 and lower.
 
 ## [0.1.0] - 2020-11-26
 ### Added